HomeWork: PerformanceTest, IndexedLists

+++ src/telran/lessons/_19/tests/PerformanceTest.java
@@ -18,7 +18,7 @@ public abstract class PerformanceTest
 			runTest();
 		}
 
-		double endTime = System.currentTimeMillis();
+		long endTime = System.currentTimeMillis();
 
 		System.out.format(
 			"Test %s [total runs: %d] performed in %f milliseconds",
